+++ Operatoren/faktionbanktauschtzentralbankgeldinbargeld.h
@@ -14,7 +14,7 @@ class FAktionBankTauschtZentralbankgeldInBargeld : public FFinanzTransaktion
 public:
 
     FAktionBankTauschtZentralbankgeldInBargeld();
-    FAktionBankTauschtZentralbankgeldInBargeld(float BETRAG, int BANKNR);
+    FAktionBankTauschtZentralbankgeldInBargeld(double BETRAG, int BANKNR);
 
     void Execute_on(FAlleDaten *AlleDaten);
 
@@ -23,7 +23,7 @@ public:
 private:
 
     // Daten
-    float Betrag;
+    double Betrag;
     int BankNr;
