+++ Operatoren/faktionbankverkauftstaatsanleiheanzbank.h
@@ -14,7 +14,7 @@ class FAktionBankVerkauftStaatsanleiheAnZBank  : public FFinanzTransaktion
 public:
 
     FAktionBankVerkauftStaatsanleiheAnZBank();
-    FAktionBankVerkauftStaatsanleiheAnZBank(float BETRAG, int BANKNR);
+    FAktionBankVerkauftStaatsanleiheAnZBank(double BETRAG, int BANKNR);
 
     void Execute_on(FAlleDaten *AlleDaten);
 
@@ -23,7 +23,7 @@ public:
 private:
 
     // Daten
-    float Betrag;
+    double Betrag;
     int BankNr;
