+++ Operatoren/faktionstaatverkauftstaatsanleihen.h
@@ -14,7 +14,7 @@ class FAktionStaatVerkauftStaatsanleihen : public FFinanzTransaktion
 public:
 
     FAktionStaatVerkauftStaatsanleihen();
-    FAktionStaatVerkauftStaatsanleihen(float BETRAG, int BANKNR);
+    FAktionStaatVerkauftStaatsanleihen(double BETRAG, int BANKNR);
 
     void Execute_on(FAlleDaten *AlleDaten);
 
@@ -23,7 +23,7 @@ public:
 private:
 
     // Daten
-    float Betrag;
+    double Betrag;
     int BankNr;
